Django/educa/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.11
django-braces==1.14.0
django-embed-video==1.4.0
django-extensions==3.1.5
django-memcache-status==2.2
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.2
Pillow==9.0.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pylibmc==1.6.1
pyOpenSSL==21.0.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uWSGI==2.0.20
Werkzeug==2.0.2
zope.interface==5.4.0
